+++ server/src/api/teams/teams.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateTeamDto } from './dto/create-team.dto';
@@ -13,21 +13,25 @@ export class TeamsService {
 
   async create(dto: CreateTeamDto) {
     const team = await this.teamRepository.create({ ...dto });
-    return this.teamRepository.save(team);
+    if (team) return this.teamRepository.save(team);
+    throw new HttpException('Команда не создана', HttpStatus.BAD_REQUEST);
   }
 
   async getAllTeams() {
     const teams = await this.teamRepository.find({ relations: ['projects'] });
-    return teams;
+    if (teams) return teams;
+    throw new HttpException('Команды не найдены', HttpStatus.NOT_FOUND);
   }
 
   async getTeamById(id: number) {
     const team = await this.teamRepository.findOneBy({ id });
-    return team;
+    if (team) return team;
+    throw new HttpException('Команда не найдена', HttpStatus.NOT_FOUND);
   }
 
   async delete(id: number): Promise<Team | undefined> {
     const team = await this.teamRepository.findOneBy({ id });
-    return this.teamRepository.remove(team);
+    if (team) return this.teamRepository.remove(team);
+    throw new HttpException('Команда не найдена', HttpStatus.NOT_FOUND);
   }
 }
